Change my email address
	* AUTHORS: Change my email address


	* tests/unpack_test.c (main): Add 64 bit tests.
	* tests/unpack_test.c (main): Use sizeof's for size.
	* tests/unpack_test.c (main): Use lg*int where we were using
	   ig*int.

	* tests/pack_test.c (test_bytes): Skip the first half
	   of 64-bit answers.  (We store the correct value as
	   32 bits.)

	* src/pack.c (gnet_vpack): Longs should get long off of stack,
not
	   int.
	* src/pack.c (gnet_vcalcsize): " (not that it matters for this
func.)

	* tests/pack_test.c (main): long tests need buffer sizeof(long).
	   (and use sizeof(unsigned int) for unsigned int tests.

+++ src/pack.c
@@ -392,8 +392,8 @@ gnet_vcalcsize (const gchar* format, va_list args)
 	case 'i':  { SIZE(int, int, gint32); 			break;  }
 	case 'I':  { SIZE(unsigned int, unsigned int, guint32); break;  }
 
-	case 'l':  { SIZE(long, int, gint32); 			break;  }
-	case 'L':  { SIZE(unsigned long, unsigned int, guint32); break;  }
+	case 'l':  { SIZE(long, long, gint32); 			break;  }
+	case 'L':  { SIZE(unsigned long, unsigned long, guint32); break;  }
 
 	case 'f':  { SIZE(float, double, float);		break;  }
 	case 'd':  { SIZE(double, double, double);		break;  }
@@ -560,8 +560,8 @@ gnet_vpack (const gchar* format, gchar* buffer, const gint length, va_list args)
 	case 'i':  { PACK2(int, int, gint32); 			break;  }
 	case 'I':  { PACK2(unsigned int, unsigned int, guint32); break;  }
 
-	case 'l':  { PACK2(long, int, gint32); 			break;  }
-	case 'L':  { PACK2(unsigned long, unsigned int, guint32); break;  }
+	case 'l':  { PACK2(long, long, gint32);			break;  }
+	case 'L':  { PACK2(unsigned long, unsigned long, guint32); break;  }
 
 	case 'f':  { PACK(float, double);			break;  }
 	case 'd':  { PACK(double, double);			break;  }
